Move parsing functions out of parse.go
